Re-add a check for the 'refresh materialized view' execution
The 'refresh materialized view' queries hang on indefinitely if executed during the tests.
Checking if we are in 'test mode', it's possible to skip their execution during the tests.

This check was already added in commit 55cd5241cfd407884ac67073c5f162f4dcadb1e7
but was removed during a merge.