[WI-827] Refactor tests number 11, 12 and 13 of the eObs API controller
These tests are all related to the 'json_take_task' route of the API controller.
Also, slightly change a method in the API controller to make it catch a more specific exception type.
N.B: other tests in the file could be failing at this point. They will be correct in the next commits.