Change group from which tests get user during setup
Using a 'default' DB (that is, a DB pre-populated with 'demo.default' data),
no user belonging to ADT group was found.
This resulted in tests failing during their setup stage.

Relying on a user belonging to the Admin group fix this problem.

N.B: despite his name, the user 'adt' present in the 'default' DB is actually an administrator,
hence changing his group in the 'demo.default' data is not a valid solution.