[WI-1240] Fix hardcoded server address
Relying on the default RouteManager's base url
raises an error when the system runs on a different server.
Fetching the correct server address from the HTTP request
(a.k.a. "on the fly") should fix the problem.

Caveat: such a fix should probably be used wherever the default
RouteManager's base url is used (currently, just in one method).
Therefore, a better long-term solution should be found and applied in future commits.