[WI-1710] Fix empty list bug in mobile controller
Add a check to ensure the list is not empty
before trying to get an element from it.
Add a safe fallback if the list is empty
(i.e. the method returns a 404 Response).

Also, add a test to catch this bug if it will show up again
in this particular mobile controller method.