Add a check to prevent exception in API setUp tests
To be run, the tests require the presence of a user in the DB.
If a user is not correctly fetched during the setUp method, the test is skipped.

This prevents an exception to be raised from inside the setUp method and
a cascade error with Odoo's database test cursor
(i.e. the cursor is left open because the exception in the setUp method
prevents the tearDown method to be executed).

Also, fix some linting.