Moved ews obs sample data to the `nh_ews` module.
The sample data is for EWS observations and so for me it made most sense for these to be made available at the same layer where ews observation themselves are first introduced as anytime from here on we may need to test with such data. Moving the sample data here made it more widely available throughout the code base.