Fixed unit tests broken due to incomplete setup.
Tests were failing after passing through the overridden complete method of `nh.clinical.patient.observation.ews` in `nh_eobs_mental_health` when a check was done on patients associated with an activity. The check did not return anything and therefore failed the calling method which assumed something would always be returned. It turned out the reason was because the activity did not have any `user_ids` associated with that it which in turn was because the users creating things were not associated with any locations.