Eobs 606 partial observation recorded in action triggered section (#85)
* - Starting to add Refused obs tests

* - Updating background.spec.js test to check value is within a range

* - Adding data point rendering, still need to add ranged graph renderinig

* - Adding ranged data points

* Amended docstrings and parameter names.

* - Amending the data points so offset properly

* - adding refused chart to desktop chart

* - Adding refused chart to report printing

* - Linting fix

* Initial implementation and unit tests for EOBS-606.

More unit tests required.

* - adding refused chart to mobile

* - Fixing nh_eobs_mobile tests by ensure referencing the new nhlib.js version

* Fixed broken XPath expressions.

* Changed test file name to match class name inside.

* Renamed test sub-package to fit convention.

The naming convention we are using is for test sub-packages to match the name of a model under test.

* Changed template and reason labels to match story.

* Logging was showing negative extra ews added.

Only the difference was required so needed to disregard the sign of the integer.

* Added implementation and unit tests, not finished yet.

* Added more fixes and unit tests.

* Fixed test.

* Fixed flake8 errors.

* Fixed tests broken by new functionality.

* Fixed Pylint errors.

* Reverted formatting to avoid tests breaking.

The formatting changed earlier resulted in beautiful soup reading whitespace around the text in text nodes. In order to fix this we need to use regex so we can match on whitespace rather than just using `self.assertEqual()` or see if there is some way beautiful soup can be configured to not read whitespace.

Also removed a redundant comment from the observation report template extension.

* Fixed flake8 and Pylint errors.

* Fixed error in template.

* Decided to create variables instead.

* Amended labels to perfectly match the story.

Labels in the story are not consistent with other labels in the system for the same partial reason but sticking to the working in the story.

* Fixed failure when no obs in report date range.

* Fixed missing tasks.

When a clinical review task is in the report date range but the observation that triggered it is not, an exception is made and the observation is included anyway. However, the triggered actions were not being included which was a mistake.

* Fixed tests and labels.

* Fixed Pylint error and test.

* Fixed test.

* Changed label after story change.

* More concise handling of missing dictionary key.

 Sometimes there is no `triggered_actions` key on the `observation` dictionary, which is fine, but must be checked for. Using `dict.get('triggered_actions', [])` is a cleaner solution than wrapping the whole loop in an `if` statement.

* - Disabling test_observation_report_rendering as it breaks Jenkins build

* - fixing 841

* - Fixing failing CI due to Pillow version not being pinned by OCA

* - Commenting why pinning dependencies in Travis.yml

* - Fixing EOBS-845