Eobs 745 clinical review fix (includes 761 and 828) (#80)
* - Adding refusal check to ensure that ir.cron record isn't created if refusal is currently underway

* - Adding refusal check to schedule notification

* - Linting Fix

* - reverting env to pool

* - Amendments after manual testing

* - Amendments after manual testing

* - Fixing tests

* - Linting Fix

* - Starting to add Refused obs tests

* - Updating background.spec.js test to check value is within a range

* - Adding data point rendering, still need to add ranged graph renderinig

* - Adding ranged data points

* - Amending the data points so offset properly

* - adding refused chart to desktop chart

* - Adding refused chart to report printing

* - Linting fix

* - adding refused chart to mobile

* - Fixing nh_eobs_mobile tests by ensure referencing the new nhlib.js version

* - Fixing issue with test

* - Restoring test after mergine 596 into develop

* Fix for EOBS-761.

When partial observations were being completed, the new observation created was not having its frequency properly set to what the partial's was (which should be whatever it still has from the last full obs and clinical risk). The frequency therefore defaulted to 15. When this new observation with incorrect frequency was completed as refused, a lookup on a dictionary is done to see if the frequency should be adjusted for refusal. Because the frequency is unexpected there is no entry for it in the dictionary and a KeyError is raised.

* Added docstrings.

* Added docstrings.

* Fixed flake8 errors.

* Moved test utility methods.

To a more generic module so they can be more widely reused.

* Fixed flake8 errors.

* Using `copy.deepcopy` instead of `dict.copy`.

Would've caused bugs because the original dictionary would have been mutated.
Also added TODO for tech debt.

* Fixed Pylint error.

* - Updating partial data references

* - Fixing OCA's Pillow dependency

* Removed usage of `current_patient_refusal_was_triggered_by` method.

* - Fixing EOBS-828

* Removed python methods replaced by SQL views.

`nh_eobs_mental_health.models.nh_clinical_patient_observation_ews.is_refusal_in_effect()` covers the same logic as these other methods.

* Added test for PME after refusal.

Also removed imports and blank line.

* Fixed flake8 errors.

* Fixed incorrect test assertion.

* - Fixing merge issue that brought back 769/770

* - Adding time between obs and transfer to ensure test doesn't fail on a technicality

* - Sorting out doc strings