Changed form description attributes.
Renamed `mandatory` attribute on form descriptions to `required` so that it matches the `required` attribute found on Odoo fields which is conceptually the same.
Added `necessary` attribute to `template.xml` so that it produces a `data-necessary` attribute on the appropriate HTML select and input elements. This will be used later for determining client-side whether the obs is partial or full.