EOBS 1117, 1260, 1233, 1237, 283, 1272, 1273, 1274, 1276, 1281, 1283 (#106)
* - Adding search domain and verification methods so that will also return PME in report if the report's date range is inside the PME's active range

* - Making sure that 'now' doesn't have seconds so every time we're using to test is 0 seconds

* - Fixing typo

* - Fixing 283: Report period now shows context away (i.e. shows BST) dates

* - Fixing 283: Report period now shows context away (i.e. shows BST) dates

* Updated rules for including PME events on report.

Now, if a stop / restart obs event falls within the date range specified for the report, any corresponding stop / restart obs event will always be included. Before, exceptions would only be made for stop obs events, now exceptions are made for both.

Also fixed a bug in the EOBS-1117 fix causing an error to be thrown when no end datetime was provided for the report.

* Added docstrings.

* Fixed flake8 errors.

* Added docstrings.

* EOBS-1259 fix.

Added a keyword argument to the `reformat_server_datetime_for_frontend` method to also convert the datetime to the client's timezone. Unfortunately the API is a little awkward in that it wants a context to be passed. The reason for this is that the `datetime_utils` model itself has no context of it's own and therefore no timezone, so the caller must pass their context which is then monkey patched onto the `datetime_utils` model.

* EOBS-1260 fix.

* Fixed flake8 errors.

* EOBS-1272, 1273, & 1274 fix.

Only added one test for the root cause of the 3 bugs. Also wanted to add some test coverage for the symptoms noted in the bugs: obs data not being displayed in mobile charts, and graphs displayed in reverse chronological order.
 The first symptom is hard to test because there is no way to know if an EWS should be included or not, we simply read it's timestamp to see if it is in the range. The ideal thing to do would be to validate that a date terminated cannot be set in the future, but such a common operations will result in a performance overhead and there may be use cases in future that require a date terminated to be set in the future. The second symptom is addressed by adding an explicit `order` parameter to the search in the `get_activity_for_spell` method which ultimately decides the order of the obs data displayed in the graphs. I am not adding a test for this as it is basic functionality of the framework's ORM that I know from experience is solid.

* Fixed pylint error.

* Getting of escalation tasks is more reliable.

Uses a reference to the observation instead of trying to determine which one it is based on the latest `date_terminated`.

* EOBS-1281 fix.

There is an override in mental health that adds EWS that are outside of the report range but do have a clinical review task that was created during the report range. The code for getting EWS within the report range correctly reformatted the datetimes but the code for the override did not. The fix was a simple matter of adding the same method call in the override as is in the parent code to reformat the datetimes. Added one unit test just to make sure that the newly added method call is always made.

* EOBS-1283 fix.

Bug was caused by the fix for EOBS-1260 which changed a regular `read` in `nh_eobs_api.get_activities_by_spell()` to the `nh_clinical_patient_observation_ews.get_formatted_obs()` which returns `None` for null values instead of `False` which is Odoo's default.

The solution was to leave the method returning `None` (as we feel this is more correct) and update a condition in the JS logic to look for any falsey value as opposed to just `false`. When the condition passes the `R` is drawn.

* Fixed another bug also caused by EOBS-1260 fix.

The EOBS-1260 fix added a `flag` to the `get_formatted_obs` method to allow the caller to choose whether they want datetimes converting to the client timezone. I did not add this flag to the overrides of the method and so when they were called with the extra keyword argument an error was raised.

Fixed by also adding the flag to overrides of the method.

* Fixed flake8 errors.