Two small fixes.
The mental health override of the transfer functionality did not call super. Mental health (incorrectly, see tech debt) depends on Food & Fluid which means that the cancel functionality in the Food & Fluid override was not called. Replaced call to `nh_clinical_api.transfer()` which a regular super call.

Cron jobs appear to be run in V7 style, passing the `cr`, `uid`, and `context` arguments. This meant that when the cron job was triggered the method call failed because too many arguments were passed. Adding the `api.model` decorator ensures that these extra arguments are accepted and packaged in the `env` before the method is called in V8 style.