[EOBS-1375] Fixed graphs not showing on reports. (#122)
* Fixed graphs not showing on reports.

In the last commit we introduced dynamic parsing of dates in the JS before they are handed over to D3. However the function assumes that the Odoo datetime format is used if none is passed in. We were returning already reformatted datetime strings from the server and then passing them into this function which was then producing invalid dates which silently failed the rendering of the charts.

 Solution was to stop the server reformatting datetimes so that they are returned in Odoo's format which the JS function expects by default. Will try to make it a rule going forward that all datetime formatting is left to the frontend.

* Fixed datetimes in report.

* Fixed datetimes again.

* Stopped redefining built-in 'format'.

* Refactored datetime handling for reports. WIP.

All datetimes are now left in Odoo server datetime format until the last moment before the template is rendered. There is a single method where all datetimes that need localising and formatting are dealt with. Other modules must add their overrides to this method for datetimes that they need reformatting.

* Fixed unit tests.

- Fixed flake8 errors.
- Changed `_localise_dict_time` to accept any falsey values and just return without exception.
- Removed test asserting that datetimes are converted in `add_extra_ews_with_clinical_review_in_datetime_range` method. The datetimes are now converted later by the `_localise_and_format_datetimes` method.
- Removed unnecessary method patch.
- Removed some conversions of datetimes for expected results in unit tests which are no longer correct because conversion is done later.

* Fixed flake8 error and unit test.