Graph datetimes not localised (#132)
* Previous fix broke datetimes, refactored.

In the previous fix the localisation occurring in `get_model_values` was removed as it was causing some datetimes in the report to be localised twice. However the graph data was relying on that localisation. The previous fix was improved encapsulation as all localisation was done in one place so it was desirable to keep that in place.

The method used for all other datetimes localised and formats them whereas the graph datetimes want localisation but not formatting. Decided to add another method called `_create_graph_data` after report data has been gathered but before the localisation of the other datetimes occurs. The method creates the JSON data and localises just those datetimes required for the graphs.

* Added docstrings.

* Removed unused imports.

* Fixed timezone failure in Travis.

Travis was failing because the client timezone was not passed to match the assertions on the datetimes occurring later on. The tests assumed BST and therefore asserted that they datetimes had been localised to BST but Travis servers are in the USA and so localised differently based on the system timezone.

Fixed the problem by adding a wrapper method for `get_report_data` that includes the additional calls to massage the data. The method needed to be created anyway for proper encapsulation and made the tests simpler as they only needed to call a single method rather than 2.