Version number visible in desktop and mobile.
Visible in desktop as an item in the menu accessible from the top right of the page.
Visible in the mobile alongside the user's name in the bottom right of the page.

Created an abstract model with a static method that gets the version number from an environment variable. The static method is called from the JS to populate the menu item on the desktop. On mobile it is called in the controller and added to the QWeb context so that it can be added directly in to the template.