Convert materialised views to regular views.
Removed materialised views to stop deadlock issues under load. This harms read performance significantly but the intention is to optimise in other areas to make up for it as we feel this is favourable over maintaining the materialised views and dealing with deadlock issues as we expect any fixes will be brittle.