Gracefully handles no environment variable.
Before we were blowing up if there was no `EOBS_VERSION` environment variable set. This is problematic because there are lots of places where we will have to ensure the environment variable is in place to avoid blow ups (Travis, every run config in IDE, Jenkins, Production). This is fine for most of them as they will want to have the environment variables but they're not really needed for Travis and local development and its quite a large maintenance overhead to add them everytime. Therefore decided to change the implementation so it only logs a warning and omits the version number from the UI.