Refactored broken unit tests.
The unit tests were using the 'requests' library and passing cookies everywhere to keep the session ID and therefore maintain authentication. This was unnecessary as Odoo's HttpCase provides an `authenticate` method that when used in combination with HttpCase's `url_open` method handles all cookies for you throughout the test case.

 Another problem was that in the setup of the test cases a nurse user was being randomly selected for login. When using an environment installed with CWP users this would intermittently fail as the CWP users do not have any passwords set and so cannot login. It is bad but we are already using demo data whether we randomly select a nurse or not so it does not seem any worse to hard core this to a known demo data nurse until we can refactor all these types of tests to not have to use demo data at all. Added a TODO linking to tech debt for this.