Removed type checking.
Realised that if I type checking in some places and not others arbitrarily. It was time for me to take a stance on the matter and apply it universally. After 10 minutes of Googling I am drawn to the opinion that type checking goes against the duck typing that Python uses, adds a lot of extra code (and therefore possible bugs), and affects performance. What we lose by not type checking is easily understandable error messages but I believe this is a small price to pay.