Changed implementation.
No longer use TransientModels with a `search_read` override. The override was too problematic as it involved reimplementing too many things such as multi-field sorting of returned records. Decided instead to create real records that are persisted using computed fields with their store attribute set to `True`. Every time the model is loaded the table will be wiped clean and models created again.

The downside of this is that location updates will not be picked up until the server is restarted or the module is re-loaded. But the upside is that it is quick and fully functional in terms of ordering, searching, and anything else you can do with Odoo's views.

Bed status and patient status fields are not stored as there is no requirement to search on these fields currently.