Moved bed management into its own menu.
Bed management was previously in the 'Open-eObs' top level menu (the horizontal menu along the top). However that menu is defined in the `nh_eobs` module which means `nh_bed_management` has to depend on it. Bed management is meant to be a potentially separate product and therefore needs to define its own menu so as not to be dependant on `nh_eobs`.