Significant changes and additions to the fix.
The backdated register activities were using current timestamp for some of the fields (such as create_date) when they should have been copying the value from the existing patient records.

The register records were now being created on admit but not when the patient was created which was still wrong. Now when on the "Create Patient Visit" page a register record is created directly because the spellboard model that powers that view now points to register rather than patient. This leaves no scope for creating a patient without a registration through the GUI. Had to create a register form view and make it look exactly like the patient form view that used to be opened before.

Made `given_name`, `family_name`, and `other_identifier` required fields on the register model to match the patient model. Updated the demo registration data to add the names.