Deleted tests that were not needed.
Decided that `test_execute` was not worth the time to implement as it was testing Odoo core code. Testing of `set_params` override would be sufficient.
Decided that testing that the `nh_clinical_frequency_setting` has certain fields was also not worth the time to implement as it is also testing Odoo core code again.