Fixed EOBS-2492.
Access to tasks was dependent on whether they were in a shift for all users but only currently nurses and HCAs can be in a shift, other user types are never part of a shift. For doctors, nurses, and HCAs their location IDs are checked to see if they are allocated to the bed they are trying to do a task for.