Changed model name to fix error.
When the patient page on the mobile is loaded it checks for all models with 'nh.clinical.patient.observation' in them, assumes they are observations, and tries to load data for them. The therapeutic level is not an observation and therefore doesn't have the expected methods for providing the expected data to the patient page which results in an error. The easiest fix was to just change the model name.